requirements.txt
aiogram==2.25.1
python-dotenv==0.19.1
